+++ src/main/java/eu/fayder/restcountries/v3/domain/BaseCountry.java
@@ -1,4 +1,4 @@
-package eu.fayder.restcountries.domain;
+package eu.fayder.restcountries.v3.domain;
 
 import java.util.List;
