+++ src/main/java/eu/fayder/restcountries/v3/domain/ResponseEntity.java
@@ -1,7 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-package eu.fayder.restcountries.domain;
+package eu.fayder.restcountries.v3.domain;
 
 public class ResponseEntity {
