+++ src/main/kotlin/eu/fayder/restcountries/v3/rest/RestCountry.kt
@@ -1,0 +1,3 @@
+package eu.fayder.restcountries.v3.rest
+
+class RestCountry
